Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/main-frame-navigated-programatically-by-subframe.html

<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpPolicyDelegateCallbacks();
    testRunner.queueLoad("resources/main-frame-with-subframe-programatically-navigates-main.html", "", true);
}

</script>
This page has does not have the "should open external URLs" flag set.<br>
It navigates to a 2nd page that does have the flag set.<br>
That 2nd page contains an iframe which programatically navigates the main frame to a 3rd page.<br>
The "should open external URLs" flag should be unset because the subframe->main frame navigation was programatic.<br>
