LayoutTests/loader/navigation-policy/should-open-external-urls/main-frame-navigated-programatically-by-subframe-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/main-frame-with-subframe-programatically-navigates-main.html, main document URL resources/main-frame-with-subframe-programatically-navigates-main.html, http method GET> is main frame - yes should open URLs externally - yes
 - decidePolicyForNavigationAction
<NSURLRequest URL programatically-navigate-to-notify-done-target-top.html, main document URL main-frame-with-subframe-programatically-navigates-main.html, http method GET> is main frame - no should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL notify-done.html, main document URL notify-done.html, http method GET> is main frame - yes should open URLs externally - no
